# src/swagger_data.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/data/courses": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "registerCoursesForDatabasePersistence",
        "operationId": "registerCoursesForDatabasePersistenceUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RegisterCoursesResponseModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/feed": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofFeed",
        "operationId": "allTimeStampListofFeedUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/feed/time/{timestamp}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "feedContentPerTimeStamp",
        "operationId": "feedContentPerTimeStampUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "description": "timestamp",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/feed/{nid}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofFeedPerNid",
        "operationId": "allTimeStampListofFeedPerNidUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/fullpost": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofAllPosts",
        "operationId": "allTimeStampListofAllPostsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/fullpost/time/{timestamp}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "postContentPerTimeStamp",
        "operationId": "postContentPerTimeStampUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "description": "timestamp",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/fullpost/{nid}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofAllPostsPerNid",
        "operationId": "allTimeStampListofAllPostsPerNidUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/stats": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofStats",
        "operationId": "allTimeStampListofStatsUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/stats/time/{timestamp}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "statsContentPerTimeStamp",
        "operationId": "statsContentPerTimeStampUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "timestamp",
            "in": "path",
            "description": "timestamp",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/data/courses/stats/{nid}": {
      "get": {
        "tags": [
          "data-crud-controller"
        ],
        "summary": "allTimeStampListofStatsPerNid",
        "operationId": "allTimeStampListofStatsPerNidUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "nid",
            "in": "path",
            "description": "nid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TimeStampListModel"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "RegisterCoursesResponseModel": {
      "type": "object",
      "properties": {
        "courseList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "timeStamp": {
          "type": "string"
        }
      }
    },
...
    "TimeStampListModel": {
      "type": "object",
      "properties": {
        "timeStampList": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "totalCount": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}